User permissions

+++ subjects.php
@@ -21,7 +21,20 @@
 
 <body class="fixed-nav sticky-footer bg-dark" id="page-top">
 <!-- Navigation-->
-<?php include "nav.php"; ?>
+<?php
+if(isset($_SESSION["username"])) {
+
+    if ($_SESSION["is_admin"] == 1) {
+        include "nav.php";
+    } elseif ($_SESSION['student_id'] != 0) {
+        include "studentnav.php";
+    } else {
+        include "professornav.php";
+    }
+}else{
+    include "guestnav.php";
+}
+?>
 <!-- -->
   <div class="content-wrapper">
     <div class="container-fluid">
@@ -45,6 +58,15 @@
                   <th>Semester</th>
                   <th>ECDL credits</th>
                   <th>Description</th>
+                    <?php
+                    if(isset($_SESSION["username"]))
+                    {
+                        if ($_SESSION["is_admin"] == 1) {
+                            echo "<th>Edit</th>";
+                            echo "<th>Delete</th>";
+                        }
+                    }
+                    ?>
                 </tr>
               </thead>
               <tbody>
@@ -56,8 +78,14 @@
                   echo "<td>" . $row['semester'] . "</td>";
                   echo "<td>" . $row['ecdl_credits'] . "</td>";
                   echo "<td>" . $row['description'] . "</td>";
-                  echo "<td><a href='edit-subject.php?id=".$row['id']."'>Edit</a></td>";
-                  echo "<td><a href='delete-subject.php?id=".$row['id']."'>Delete</a></td>";
+                  if(isset($_SESSION["username"]))
+                  {
+                      if ($_SESSION["is_admin"] == 1)
+                      {
+                          echo "<td><a href='edit-subject.php?id=" . $row['id'] . "'>Edit</a></td>";
+                          echo "<td><a href='delete-subject.php?id=" . $row['id'] . "'>Delete</a></td>";
+                      }
+                  }
                   echo "</tr>";
               }
               mysqli_close($conn);
